refactor file storage model
