+++ src/sts/models/file_storage/__init__.py
@@ -1,0 +1,23 @@
+from sts.models.file_storage.image_data import ImageData
+from sts.models.file_storage.storage_item import StorageFileItem, StorageResponse
+from sts.models.file_storage.scan_result import (
+    ScanResultNotFound,
+    ScanResultUseSourceFile,
+    ScanResultFileFound,
+    ScanResultCreateNew,
+    ScanResult
+)
+
+__all__ = [
+    # image data
+    "ImageData",
+    # storage item
+    "StorageFileItem",
+    "StorageResponse",
+    # scan result
+    "ScanResult",
+    "ScanResultNotFound",
+    "ScanResultUseSourceFile",
+    "ScanResultFileFound",
+    "ScanResultCreateNew",
+]
