+++ src/sts/models/file_storage/image_data.py
@@ -1,0 +1,11 @@
+from io import BytesIO
+
+from pydantic import ConfigDict
+from pydantic.dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True, config=ConfigDict(arbitrary_types_allowed=True))
+class ImageData:
+    content_type: str
+    error: Exception | None = None
+    data: BytesIO | None = None
